app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
databricks-sdk>=0.30.0
databricks-sql-connector>=3.1.0
httpx>=0.27.0
pydantic>=2.0.0
